+++ src/AtomUI.Desktop.Controls.DataGrid/EventArgs/DataGridColumnFilterEventArgs.cs
@@ -2,11 +2,11 @@ namespace AtomUI.Desktop.Controls;
 
 public class DataGridColumnFilterEventArgs : EventArgs
 {
-    public List<string> FilterValues { get; }
+    public List<object> FilterValues { get; }
 
     public DataGridColumn Column { get; }
 
-    public DataGridColumnFilterEventArgs(DataGridColumn column, List<string> filterValues)
+    public DataGridColumnFilterEventArgs(DataGridColumn column, List<object> filterValues)
     {
         Column       = column;
         FilterValues = filterValues;